predator-v4/requirements.txt
ccxt>=4.4.0
aiohttp>=3.9.0
websockets>=13.0
httpx>=0.27.0
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.5.0
joblib>=1.4.0
orjson>=3.9.0
loguru>=0.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
uvloop>=0.19.0
